ILM: enable the readonly action in the cold phase (#69024)
This enables the readonly ILM action in the cold phase.

(cherry picked from commit 9da3a6ee8d74baaca70d3706082c3613b21aeadf)
